Improving folder and file structure in project
