+++ app/weather-ui/src/backend/server.ts
@@ -1,19 +1,15 @@
 import path from "path";
 import formBody from "@fastify/formbody"; // eslint-disable-line
 import staticFiles from "@fastify/static"; // eslint-disable-line
-import axios from "axios"; // eslint-disable-line
 import dotenv from "dotenv"; // eslint-disable-line
 import { fastify as fastify_fastify } from "fastify"; // eslint-disable-line
 import nunjucks from "nunjucks"; // eslint-disable-line
 import { z } from "zod"; // eslint-disable-line
 import { fetchLocationData } from "./location";
 import { fetchWeatherData } from "./weatherapi";
+import { weatherCodeToImage } from "./constants";
 
 dotenv.config();
-
-const WEATHER_API_URL = "https://api.open-meteo.com/v1/forecast";
-const GEOCODE_API_URL = "https://geocode.maps.co/search";
-const HTTP_CLIENT = axios;
 
 const environment = process.env.NODE_ENV;
 const templates = new nunjucks.Environment(new nunjucks.FileSystemLoader("src/backend/templates"));
@@ -21,7 +17,6 @@ const templates = new nunjucks.Environment(new nunjucks.FileSystemLoader("src/ba
 const server = fastify_fastify({
     logger: true,
 });
-
 
 // setup middlewares
 {
@@ -34,50 +29,15 @@ const server = fastify_fastify({
     });
 }
 
-const weatherCodeToImage = (code: number): string => {
-  switch (code) {
-    case 0: return "/static/img/clear.svg";
-    case 1: return "/static/img/clear.svg";
-    case 2: return "/static/img/cloudy.svg";
-    case 3: return "/static/img/overcast.svg";
-    case 45: return "/static/img/fog.svg";
-    case 48: return "/static/img/fog.svg";
-    case 51: return "/static/img/drizzle.svg";
-    case 53: return "/static/img/drizzle.svg";
-    case 55: return "/static/img/drizzle.svg";
-    case 56: return "/static/img/drizzle.svg";
-    case 57: return "/static/img/drizzle.svg";
-    case 61: return "/static/img/rain.svg";
-    case 63: return "/static/img/rain.svg";
-    case 65: return "/static/img/rain.svg";
-    case 66: return "/static/img/rain.svg";
-    case 67: return "/static/img/rain.svg";
-    case 71: return "/static/img/snow.svg";
-    case 73: return "/static/img/snow.svg";
-    case 75: return "/static/img/snow.svg";
-    case 77: return "/static/img/snow.svg";
-    case 80: return "/static/img/rain.svg";
-    case 81: return "/static/img/rain.svg";
-    case 82: return "/static/img/rain.svg";
-    case 85: return "/static/img/snow.svg";
-    case 86: return "/static/img/snow.svg";
-    case 95: return "/static/img/thunderstorm.svg";
-    case 96: return "/static/img/thunderstorm.svg";
-    case 99: return "/static/img/thunderstorm.svg";
-    default: return "/static/img/info.svg";
-  }
-};
-
-const locationSchema = z.object({
-    location: z.string(),
-});
-
 server.get("/", async (request, reply) => {
+    const locationSchema = z.object({
+        location: z.string(),
+    });
     const queryParams = request.query;
     try {
         const { location } = locationSchema.parse(queryParams);
-        const locationInfo = await fetchLocationData(HTTP_CLIENT, GEOCODE_API_URL, location); // eslint-disable-line
-        const weatherInfo = await fetchWeatherData(HTTP_CLIENT, WEATHER_API_URL, locationInfo.lat, locationInfo.lon); // eslint-disable-line
+        const locationInfo = await fetchLocationData(location); // eslint-disable-line
+        const weatherInfo = await fetchWeatherData(locationInfo.lat, locationInfo.lon); // eslint-disable-line
 
         const rendered = templates.render("weather.njk", {
             environment,
